reactored functions
