+++ src/pages/alle-oppgaver.tsx
@@ -9,9 +9,10 @@ import moment from 'moment';
 import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
+import { useState } from 'react';
 import safeJsonStringify from 'safe-json-stringify';
 import { IEmployeeTask } from 'utils/types';
-import { splitIntoTimeSections } from 'utils/utils';
+import { searchTask, splitIntoTimeSections } from 'utils/utils';
 
 const useStyles = makeStyles({
   root: {
@@ -103,6 +104,13 @@ const MyTasks = ({ myTasks }: InferGetServerSidePropsType<typeof getServerSidePr
 
   const timeSections: TimeSectionType[] = splitIntoTimeSections(myTasks);
 
+  const [searchResults, setSearchResults] = useState([]);
+
+  const search = (text: string) => {
+    const result = searchTask(text, timeSections);
+    setSearchResults(result);
+  };
+
   return (
     <>
       <Head>
@@ -115,10 +123,14 @@ const MyTasks = ({ myTasks }: InferGetServerSidePropsType<typeof getServerSidePr
           </Typo>
           <Typo className={classes.template_title}>{completed.toString() === 'true' ? 'Fullførte' : 'Aktive'} oppgaver</Typo>
         </div>
-        <SearchFilter />
+        <SearchFilter search={search} />
         <div>
           {timeSections.length === 0 ? (
             <Typo>Ingen oppgaver</Typo>
+          ) : searchResults.length > 0 ? (
+            searchResults.map((section: TimeSectionType, index: number) => {
+              return <TimeSection first={index === 0} key={index} section={section} />;
+            })
           ) : (
             timeSections.map((section: TimeSectionType, index: number) => <TimeSection first={index === 0} key={index} section={section} />)
           )}
